Update Drawer tests
