+++ services/ui-src/src/components/drawers/ReportDrawerDetails.test.tsx
@@ -2,46 +2,29 @@ import { render, screen } from "@testing-library/react";
 import { axe } from "jest-axe";
 //components
 import { ReportDrawerDetails } from "components";
-
-const mockAccessMeasuresDrawerDetails = {
-  standardDescription: "mock description",
-  category: "mock category",
-};
+import {
+  mockUnfinishedAccessMeasuresFormattedEntityData,
+  mockUnfinishedQualityMeasuresFormattedEntityData,
+  mockUnfinishedSanctionsFormattedEntityData,
+} from "utils/testing/setupJest";
 
 const ReportDrawerDetailsAccessMeasuresComponent = (
   <ReportDrawerDetails
-    drawerDetails={mockAccessMeasuresDrawerDetails}
+    drawerDetails={mockUnfinishedAccessMeasuresFormattedEntityData}
     entityType="accessMeasures"
   />
 );
 
-const mockSanctionsDrawerDetails = {
-  interventionType: "mock intervention type",
-  interventionTopic: "mock intervention topic",
-  planName: "mock plan name",
-  interventionReason: "mock intervention reason",
-};
-
 const ReportDrawerDetailsSanctionsComponent = (
   <ReportDrawerDetails
-    drawerDetails={mockSanctionsDrawerDetails}
+    drawerDetails={mockUnfinishedSanctionsFormattedEntityData}
     entityType="sanctions"
   />
 );
 
-const mockQualityMeasuresDrawerDetails = {
-  domain: "mock domain",
-  name: "mock quality measure name",
-  nqfNumber: "mock nqf number",
-  reportingRateType: "mock reporting rate type",
-  set: "mock measure set",
-  reportingPeriod: "mock reporting period",
-  description: "mock description",
-};
-
 const ReportDrawerDetailsQualityMeasuresComponent = (
   <ReportDrawerDetails
-    drawerDetails={mockQualityMeasuresDrawerDetails}
+    drawerDetails={mockUnfinishedQualityMeasuresFormattedEntityData}
     entityType="qualityMeasures"
   />
 );
@@ -54,24 +37,26 @@ describe("Test ReportDrawerDetails renders given text", () => {
   it("Should render access measures text provided in drawerDetails", async () => {
     render(ReportDrawerDetailsAccessMeasuresComponent);
     expect(
-      screen.getByText(mockAccessMeasuresDrawerDetails.category)
+      screen.getByText(mockUnfinishedAccessMeasuresFormattedEntityData.category)
     ).toBeVisible();
   });
 
   it("Should render sanctions text provided in drawerDetails", async () => {
     render(ReportDrawerDetailsSanctionsComponent);
     expect(
-      screen.getByText(mockSanctionsDrawerDetails.interventionTopic)
+      screen.getByText(
+        mockUnfinishedSanctionsFormattedEntityData.interventionTopic
+      )
     ).toBeVisible();
   });
 
   it("Should render quality measures text provided in drawerDetails", async () => {
     render(ReportDrawerDetailsQualityMeasuresComponent);
     expect(
-      screen.getByText(mockQualityMeasuresDrawerDetails.domain)
+      screen.getByText(mockUnfinishedQualityMeasuresFormattedEntityData.domain)
     ).toBeVisible();
     expect(
-      screen.getByText(mockQualityMeasuresDrawerDetails.name)
+      screen.getByText(mockUnfinishedQualityMeasuresFormattedEntityData.name)
     ).toBeVisible();
   });
 });
