+++ services/ui-src/src/utils/testing/mockAsset.ts
@@ -1,1 +1,24 @@
+jest.mock("aws-amplify", () => ({
+  Auth: {
+    currentSession: jest.fn().mockReturnValue({
+      getIdToken: () => ({
+        getJwtToken: () => "eyJLongToken",
+      }),
+    }),
+    currentAuthenticatedUser: () => {},
+    configure: () => {},
+    signOut: async () => {},
+    federatedSignIn: () => {},
+  },
+  API: {
+    get: () => {},
+    post: () => {},
+    put: () => {},
+    del: () => {},
+    configure: () => {},
+  },
+  Hub: {
+    listen: jest.fn(),
+  },
+}));
 export default "test-file-stub";
