+++ services/ui-src/src/utils/testing/mockBanner.tsx
@@ -1,0 +1,9 @@
+import { bannerId } from "../../constants";
+
+export const mockBannerData = {
+  key: bannerId,
+  title: "Yes here I am, a banner",
+  description: "I have a description too thank you very much",
+  startDate: 1640995200000, // 1/1/2022 00:00:00 UTC
+  endDate: 1672531199000, // 12/31/2022 23:59:59 UTC
+};
