+++ services/ui-src/src/utils/testing/mockForm.tsx
@@ -1,0 +1,238 @@
+export const mockFormField = {
+  id: "mock-text-field",
+  type: "text",
+  validation: "text",
+  props: {
+    label: "mock text field",
+  },
+};
+
+export const mockRepeatedFormField = {
+  id: "mock-text-field",
+  type: "text",
+  validation: "text",
+  repeat: "plans",
+  props: {
+    label: "mock text field",
+  },
+};
+
+export const mockModalFormField = {
+  id: "mock-modal-text-field",
+  type: "text",
+  validation: "text",
+  props: {
+    label: "mock modal text field",
+  },
+};
+
+export const mockModalOverlayFormField = {
+  id: "mock-modal-overlay-text-field",
+  type: "text",
+  validation: "text",
+  props: {
+    label: "mock modal overlay text field",
+  },
+};
+
+export const mockDrawerFormField = {
+  id: "mock-drawer-text-field",
+  type: "text",
+  validation: "text",
+  props: {
+    label: "mock drawer text field",
+  },
+};
+
+export const mockNestedFormField = {
+  id: "mock-nested-field",
+  type: "radio",
+  validation: "radio",
+  props: {
+    label: "mock radio field",
+    choices: [
+      { id: "option1uuid", label: "option 1" },
+      { id: "option2uuid", label: "option 2" },
+      {
+        id: "option3uuid",
+        label: "option 3",
+        children: [mockFormField],
+      },
+    ],
+  },
+};
+
+export const mockPlanField = {
+  id: "plans",
+  type: "dynamic",
+  validation: "dynamic",
+  props: {
+    label: "Plan name",
+  },
+};
+
+export const mockForm = {
+  id: "mock-form-id",
+  fields: [mockFormField],
+};
+
+export const mockModalForm = {
+  id: "mock-modal-form-id",
+  fields: [mockModalFormField],
+};
+
+export const mockDrawerForm = {
+  id: "mock-drawer-form-id",
+  fields: [mockDrawerFormField],
+};
+
+export const mockEmptyDrawerForm = {
+  id: "mock-drawer-form-id",
+  fields: [],
+};
+
+export const mockModalOverlayForm = {
+  id: "mock-modal-overlay-form-id",
+  fields: [
+    mockModalOverlayFormField,
+    { ...mockModalOverlayFormField, id: "with-label", props: { label: "1" } },
+    mockNestedFormField,
+  ],
+};
+
+export const mockPlanFilledForm = {
+  id: "mock-form-id",
+  fields: [mockPlanField],
+};
+
+export const mockNestedForm = {
+  id: "mock-nested-form-id",
+  fields: [mockNestedFormField],
+};
+
+export const mockDynamicForm = {
+  id: "mock-dynamic-form-id",
+  fields: [mockPlanField],
+};
+
+export const mockLinksForm = {};
+
+export const mockVerbiageIntro = {
+  section: "mock section",
+  subsection: "mock subsection",
+};
+
+export const mockStandardReportPageJson = {
+  name: "mock-route-1",
+  path: "/mock/mock-route-1",
+  pageType: "standard",
+  verbiage: {
+    intro: mockVerbiageIntro,
+  },
+  form: mockForm,
+};
+
+export const mockDynamicReportPageJson = {
+  name: "mock-route-1",
+  path: "/mock/mock-route-1",
+  pageType: "standard",
+  verbiage: {
+    intro: mockVerbiageIntro,
+  },
+  form: mockDynamicForm,
+};
+
+export const mockDrawerReportPageJson = {
+  name: "mock-route-2a",
+  path: "/mock/mock-route-2a",
+  pageType: "drawer",
+  entityType: "plans",
+  verbiage: {
+    intro: mockVerbiageIntro,
+    dashboardTitle: "Mock dashboard title",
+    drawerTitle: "Mock drawer title",
+  },
+  drawerForm: mockDrawerForm,
+};
+
+export const mockNestedReportPageJson = {
+  name: "mock-route-2a",
+  path: "/mock/mock-route-2a",
+  pageType: "drawer",
+  entityType: "plans",
+  verbiage: {
+    intro: mockVerbiageIntro,
+    dashboardTitle: "Mock dashboard title",
+    drawerTitle: "Mock drawer title",
+  },
+  drawerForm: mockNestedForm,
+};
+
+export const mockModalDrawerReportPageVerbiage = {
+  intro: mockVerbiageIntro,
+  dashboardTitle: "Mock dashboard title",
+  addEntityButtonText: "Mock add entity button text",
+  editEntityButtonText: "Mock edit entity button text",
+  addEditModalAddTitle: "Mock add/edit entity modal add title",
+  addEditModalEditTitle: "Mock add/edit entity modal edit title",
+  addEditModalMessage: "Mock add/edit entity modal message",
+  deleteEntityButtonAltText: "Mock delete entity button alt text",
+  deleteModalTitle: "Mock delete modal title",
+  deleteModalConfirmButtonText: "Mock delete modal confirm button text",
+  deleteModalWarning: "Mock delete modal warning",
+  entityUnfinishedMessage: "Mock entity unfinished messsage",
+  enterEntityDetailsButtonText: "Mock enter entity details button text",
+  editEntityDetailsButtonText: "Mock edit entity details button text",
+  drawerTitle: "Mock drawer title",
+  drawerNoFormMessage: "Mock no form fields here",
+};
+
+export const mockModalOverlayReportPageVerbiage = {
+  intro: mockVerbiageIntro,
+  dashboardTitle: "Mock dashboard title",
+  addEditModalHint: "Mock modal hint",
+  countEntitiesInTitle: true,
+  tableHeader: "Mock table header",
+  addEntityButtonText: "Mock add entity button text",
+  emptyDashboardText: "Mock empty dashboard text",
+  editEntityButtonText: "Mock edit entity button text",
+  deleteModalTitle: "Mock delete modal title",
+  deleteModalConfirmButtonText: "Mock delete modal confirm button text",
+  deleteModalWarning: "Mock delete modal warning",
+  enterReportText: "Mock enter report text",
+};
+
+export const mockModalDrawerReportPageJson = {
+  name: "mock-route-2b",
+  path: "/mock/mock-route-2b",
+  pageType: "modalDrawer",
+  entityType: "accessMeasures",
+  verbiage: mockModalDrawerReportPageVerbiage,
+  modalForm: mockModalForm,
+  drawerForm: mockDrawerForm,
+};
+
+export const mockModalOverlayReportPageJson = {
+  name: "mock-route-2c",
+  path: "/mock/mock-route-2c",
+  pageType: "modalOverlay",
+  entityType: "program",
+  verbiage: mockModalOverlayReportPageVerbiage,
+  modalForm: mockModalOverlayForm,
+};
+
+export const mockModalOverlayReportPageWithOverlayJson = {
+  name: "mock-route-2c",
+  path: "/mock/mock-route-2c",
+  pageType: "modalOverlay",
+  entityType: "program",
+  verbiage: mockModalOverlayReportPageVerbiage,
+  modalForm: mockModalOverlayForm,
+  overlayForm: mockModalOverlayForm,
+};
+
+export const mockReviewSubmitPageJson = {
+  name: "mock-route-3",
+  path: "/mock/mock-review-and-submit",
+  pageType: "reviewSubmit",
+};
