+++ services/ui-src/src/utils/testing/mockRouter.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { BrowserRouter as Router } from "react-router-dom";
+
+export const RouterWrappedComponent: React.FC = ({ children }) => (
+  <Router>{children}</Router>
+);
